packages/alpaca_client/pyproject.toml
[project]
name = "kt-alpaca"
version = "0.1.0"
description = "Alpaca API client for trading and market data"
requires-python = ">=3.12"
dependencies = [
    "kt-shared",
    "alpaca-py>=0.31",
    "httpx>=0.27",
    "pandas>=2.0",
]
